fixes some bugs

+++ src/main/java/edu/wpi/N/views/admin/AddDoctorController.java
@@ -7,6 +7,7 @@ import edu.wpi.N.database.DBException;
 import edu.wpi.N.database.DoctorDB;
 import edu.wpi.N.database.ServiceDB;
 import edu.wpi.N.entities.DbNode;
+import edu.wpi.N.entities.employees.Doctor;
 import java.util.LinkedList;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -87,50 +88,49 @@ public class AddDoctorController {
   }
 
   public void addOffice() throws DBException {
-    int currentSelection = lst_docoffice.getSelectionModel().getSelectedIndex();
+    try {
+      int currentSelection = lst_docoffice.getSelectionModel().getSelectedIndex();
 
-    if (currentSelection == -1) {
-      Alert acceptReq = new Alert(Alert.AlertType.ERROR);
-      acceptReq.setContentText("Invalid / No Location Selected");
-      acceptReq.show();
-
-      return;
-    }
-
-    if (!ServiceDB.getEmployee(Integer.parseInt(txtf_docid.getText()))
-        .getServiceType()
-        .equals("Medicine")) {
-
-      Alert invalidID = new Alert(Alert.AlertType.ERROR);
-      invalidID.setContentText("Employee isn't a doctor.");
-      invalidID.show();
-
-      return;
-    }
-
-    DbNode addOfficeNode = fuzzySearchNodeList.get(currentSelection);
-
-    if (txtf_docid.getText().equals("")) {
-      Alert invalidID = new Alert(Alert.AlertType.ERROR);
-      invalidID.setContentText("Invalid ID");
-      invalidID.show();
-
-      return;
-    }
-
-    int doctorID = Integer.parseInt(txtf_docid.getText());
-
-    for (DbNode node : DoctorDB.getDoctor(doctorID).getLoc()) {
-      if (node.equals(addOfficeNode)) {
+      if (currentSelection == -1) {
         Alert acceptReq = new Alert(Alert.AlertType.ERROR);
-        acceptReq.setContentText("Doctor already works there");
+        acceptReq.setContentText("Invalid / No Location Selected");
         acceptReq.show();
 
         return;
       }
-    }
 
-    try {
+      if (!ServiceDB.getEmployee(Integer.parseInt(txtf_docid.getText()))
+          .getServiceType()
+          .equals("Medicine")) {
+
+        Alert invalidID = new Alert(Alert.AlertType.ERROR);
+        invalidID.setContentText("Employee isn't a doctor.");
+        invalidID.show();
+
+        return;
+      }
+
+      DbNode addOfficeNode = fuzzySearchNodeList.get(currentSelection);
+
+      if (txtf_docid.getText().equals("")) {
+        Alert invalidID = new Alert(Alert.AlertType.ERROR);
+        invalidID.setContentText("Invalid ID");
+        invalidID.show();
+
+        return;
+      }
+
+      int doctorID = Integer.parseInt(txtf_docid.getText());
+
+      for (DbNode node : DoctorDB.getDoctor(doctorID).getLoc()) {
+        if (node.equals(addOfficeNode)) {
+          Alert acceptReq = new Alert(Alert.AlertType.ERROR);
+          acceptReq.setContentText("Doctor already works there");
+          acceptReq.show();
+
+          return;
+        }
+      }
 
       DoctorDB.addOffice(doctorID, addOfficeNode);
 
@@ -142,11 +142,6 @@ public class AddDoctorController {
       Alert errorAlert = new Alert(Alert.AlertType.ERROR);
       errorAlert.setContentText(e.getMessage());
       errorAlert.show();
-    }
-
-    for (DbNode node : DoctorDB.getDoctor(doctorID).getLoc()) {
-      System.out.println("Node: " + node.getLongName());
-      System.out.println("Given: " + addOfficeNode.getLongName());
     }
 
     txtf_docid.clear();
@@ -167,6 +162,13 @@ public class AddDoctorController {
       invalidID.show();
 
       return;
+    }
+
+    if (ServiceDB.getEmployee(Integer.parseInt(txtf_docid.getText())) == null) {
+
+      Alert invalidID = new Alert(Alert.AlertType.ERROR);
+      invalidID.setContentText("Employee doesnt exist");
+      invalidID.show();
     }
 
     if (!ServiceDB.getEmployee(Integer.parseInt(txtf_docid.getText()))
@@ -196,14 +198,21 @@ public class AddDoctorController {
 
     for (DbNode node : DoctorDB.getDoctor(docid).getLoc()) {
       if (node.getLongName().equals(removeOfficeNode.getLongName())) {
+        for (Doctor docs : DoctorDB.getDoctors()) {
+          if (docid == docs.getID()) {
+            DoctorDB.removeOffice(Integer.parseInt(txtf_docid.getText()), removeOfficeNode);
 
-        DoctorDB.removeOffice(Integer.parseInt(txtf_docid.getText()), removeOfficeNode);
+            Alert acceptReq = new Alert(Alert.AlertType.CONFIRMATION);
+            acceptReq.setContentText("Office " + removeOfficeNode.getLongName() + " removed.");
+            acceptReq.show();
 
-        Alert acceptReq = new Alert(Alert.AlertType.CONFIRMATION);
-        acceptReq.setContentText("Office " + removeOfficeNode.getLongName() + " removed.");
-        acceptReq.show();
+            return;
+          }
 
-        return;
+          Alert acceptReq = new Alert(Alert.AlertType.ERROR);
+          acceptReq.setContentText("Doctor with that ID doesnt exist");
+          acceptReq.show();
+        }
       }
     }
